+++ reptar/extractors/__init__.py
@@ -2,7 +2,8 @@ from .extractor import extractor
 from .orca_extractor import extractorORCA
 from .xtb_extractor import extractorXTB
 from .ase_extractor import extractorASE
+from .crest_extractor import extractorCREST
 
 __all__ = [
-    'extractor', 'extractorORCA', 'extractorXTB', 'extractorASE'
+    'extractor', 'extractorORCA', 'extractorXTB', 'extractorASE', 'extractorCREST'
 ]