+++ reptar/parsers/__init__.py
@@ -2,7 +2,8 @@ from .parser import parser
 from .orca_parser import parserORCA
 from .xtb_parser import parserXTB
 from .ase_parser import parserASE
+from .crest_parser import parserCREST
 
 __all__ = [
-    'parser', 'parserORCA', 'parserXTB', 'parserASE'
+    'parser', 'parserORCA', 'parserXTB', 'parserASE', 'parserCREST'
 ]